Solved the issue of Navigation Graph and succesfully implemented he Navigation View along with the menu ids of Bottom Navigation Menu
