+++ app/src/main/java/com/example/visuallithuanian/UserFragment.kt
@@ -1,59 +1,52 @@
 package com.example.visuallithuanian
 
+import android.annotation.SuppressLint
+import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.GridView
+import com.example.visuallithuanian.adapter.LanguageAdapter
+import com.example.visuallithuanian.databinding.FragmentUserBinding
+import com.example.visuallithuanian.model.LanguageModel
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
+import com.google.android.gms.auth.api.signin.GoogleSignInClient
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.common.api.ApiException
+import com.google.android.gms.tasks.Task
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [UserFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class UserFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_user, container, false)
+        val view = inflater.inflate(R.layout.fragment_user, container, false)
+        val gridView = view.findViewById<GridView>(R.id.mainSpecimens)
+
+        val languageModelArrayList:ArrayList<LanguageModel> = ArrayList()
+        languageModelArrayList.add(LanguageModel("Vaisiai",R.drawable.fruits,"Fruits"))
+        languageModelArrayList.add(LanguageModel("Gėlės",R.drawable.flowers,"Flowers"))
+        languageModelArrayList.add(LanguageModel("Daržovės",R.drawable.vegetable,"Vegetables"))
+        languageModelArrayList.add(LanguageModel("Architektūros",R.drawable.campus,"Architectures"))
+        languageModelArrayList.add(LanguageModel("Saulės Sistema",R.drawable.universe,"Solar System"))
+        languageModelArrayList.add(LanguageModel("Sezonus",R.drawable.season,"Seasons"))
+        languageModelArrayList.add(LanguageModel("Kompasas",R.drawable.compass,"Compass"))
+        languageModelArrayList.add(LanguageModel("Gyvūnai",R.drawable.animals,"Animals"))
+        languageModelArrayList.add(LanguageModel("Paukščiai",R.drawable.birds,"Birds"))
+        languageModelArrayList.add(LanguageModel("Sporto",R.drawable.sports,"Sports"))
+
+        val languageAdapter = context?.let { LanguageAdapter(it,languageModelArrayList) }
+        gridView.adapter=languageAdapter
+        gridView.isNestedScrollingEnabled = true
+        return view
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment UserFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            UserFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
-    }
+
+
 }
