+++ app/src/main/res/navigation/nav_graph.xml
@@ -4,20 +4,18 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
     app:startDestination="@id/userFragment">
-
     <fragment
         android:id="@+id/userFragment"
         android:name="com.example.visuallithuanian.UserFragment"
         android:label="fragment_user"
-        tools:layout="@layout/fragment_user" />
+        tools:layout="@layout/fragment_user" >
+        <action
+            android:id="@+id/action_userFragment_to_locationFragment"
+            app:destination="@id/locationFragment" />
+    </fragment>
     <fragment
         android:id="@+id/locationFragment"
         android:name="com.example.visuallithuanian.LocationFragment"
         android:label="fragment_location"
         tools:layout="@layout/fragment_location" />
-    <activity
-        android:id="@+id/first_screen"
-        android:name="com.example.visuallithuanian.ui.activities.FirstScreen"
-        android:label="first_screen"
-        tools:layout="@layout/first_screen" />
 </navigation>
